Extract ParseUtils
